requirements.txt
alembic==1.4.2
Babel==2.8.0
bleach==2.0.0
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
dominate==2.5.1
elasticsearch==7.9.1
email-validator==1.1.1
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
guess-language-spirit==0.5.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
Markdown==2.6.9
MarkupSafe==1.1.1
pathlib==1.0.1
Pootle==2.8.2
pycountry==17.5.14
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-Levenshtein==0.12.0
pytz==2020.1
requests==2.24.0
rq==0.8.2
scandir==1.5
six==1.15.0
SQLAlchemy==1.3.17
stemming==1.0.1
translate-toolkit==2.2.5
urllib3==1.25.10
uuid==1.30
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
